Linux port: Comprehensive porting work - UI, campaign, 3D rendering, input
Major subsystems ported from Windows/DirectX to Linux/SDL2/OpenGL:

- UI95 widget toolkit: Resource loading (.idx/.rsc), case-insensitive file
  lookup, 32-bit binary compatibility for image/sound/flat headers, mouse
  event routing from SDL to UI handler, coordinate scaling

- Campaign system: LZSS decompression, 32/64-bit binary compatibility for
  save files (long -> int32_t/uint32_t), VU message dispatch during
  deaggregation, mode transition fixes in sim loop

- 3D rendering pipeline: D3D7-to-OpenGL compatibility layer, XYZRHW
  pre-transformed vertex handling, texture loading with case-insensitive
  paths, 64-bit pointer truncation fixes throughout graphics allocator,
  DXContext, CDXEngine, cockpit rendering

- Input system: SDL-to-DIK scancode translation, SDL joystick support with
  axis mapping (roll/pitch/yaw/throttle), POV hat, hot-plugging

- Audio: OpenAL-backed DirectSound implementation, WAV loading fixes

- Thread safety: GL context transfer between UI and sim threads, volatile
  currentMode for cross-thread visibility, race condition fixes in
  UI_Cleanup vs OTWDriver.Enter

- Stability: NULL safety checks in cockpit DisplayBlit3D, VU thread,
  graphics pipeline; snprintf replacements; debug macro system

+++ src/campaign/camptask/atm.cpp
@@ -12,7 +12,7 @@
 #include <time.h>
 #include "campmap.h"
 #include "CmpGlobl.h"
-#include "ListADT.h"
+#include "listadt.h"
 #include "vutypes.h"
 #include "Objectiv.h"
 #include "Find.h"
@@ -1609,7 +1609,7 @@ Squadron AirTaskingManagerClass::FindBestAir(MissionRequest mis, GridIndex bx, G
                 // RV - Biker - Reduce loiter time for short range AC
                 if (d + db > sq->GetUnitRange())
                 {
-                    MissionData[mis->mission].loitertime = min(45, MissionData[mis->mission].loitertime);
+                    MissionData[mis->mission].loitertime = min((short)45, MissionData[mis->mission].loitertime);
                     db = (MissionData[mis->mission].loitertime * speed) / 60.0F;
 
                     if (d + db > sq->GetUnitRange())
